mend

+++ src/use-cases/province/destroy-province.ts
@@ -1,4 +1,5 @@
 import { ProvincesRepository } from "@/repositories/province-repository"
+import { ProvinceNotFoundError } from "../errors/province-not-found"
 
 interface DestroyProvinceUseCaseRequest {
   id: number
@@ -10,6 +11,11 @@ export class DestroyProvinceUseCase {
   async execute({
     id,
   }: DestroyProvinceUseCaseRequest): Promise<Boolean> {
+
+    let findProvince = await this.provincesRepository.findById(id)
+    if(!findProvince){
+      throw new ProvinceNotFoundError();
+    }
     return await this.provincesRepository.destroy(
       id,
     )
